requirements.txt
hrequests
requests

pytest
pytest-typed-schema-shot